Yet another bug fix to prog/desc indices...
This time for the case when a galaxy merges into another galaxy that has
zero stellar mass.  Now we don't bother setting the next progenitor
indices for the central (zero stellar mass) galaxy.  This means that we
are loosing part of the galaxy history if we ever try and trace it back
in the output, but likely just for fluff around the resolution limit
anyway.