Try to use FirstOccupiedHalo to allow empty haloes.
This implementation is overly complicated.  FirstOccupiedHalo should
just be initialised to NULL for all FOFGroups and then an extra loop
used to calculate its value immediately prior to calling
evolve_galaxies() in dracarys.c.  I'm going to leave this for a future
exercise after the first paper is done though as I don't have time right
now!

In fact - a good week of cleaning up and removing the stratified layers
of checks and calculations is definitely called for when I have time!