Code now builds but is untested
There are a few galaxy and halo properties I removed for the time being
as they are not currently being used and it reduced the amount of
decisions I needed to make.  Even if the code runs at this point though,
there are a lot of hasty and not well motivated decisions I have made.