Complete comparison with old Tiamat version.
There is one issue remaining, but I think the new version is actually
doing a better job than the old one!

For galaxy `ID=140017467636` at `snapshot=19`, the original model hadn’t
fully connected the pointers between all participants of a three way
merger before calculating the merger time of this galaxy.  The result is
that the wrong subhalo was used as the `mother` in the timescale
calculation.  In the new version of the model, all of the pointers are
connected first, and the correct mother is identified.

Unfortunately, there is no way for me to correctly “break” the new
version of the model such that I get the same result as the old version
for cases like these.  This means my comparison has to stop here… . The
model does make it to to snapshot 18 (inclusive) and gives the exact
same result (using `numpy.isclose`) as the original version.  I am
therefore confident now that things are working correctly…